+++ frontend/src/api/Authentication.js
@@ -4,20 +4,24 @@ import serverURL from './config';
 import 'firebase/compat/auth';
 import 'firebase/compat/firestore';
 
-export function signInWithGoogle(setUserID) {
+export function signInWithGoogle(setJwtToken, rawAccessToken, accessToken, setErrorMessage, forceRecreate) {
     const provider = new firebase.auth.GoogleAuthProvider();
   
     // Sign in
     firebase.auth().signInWithPopup(provider)
-        .then((result) => {
+        .then(async (result) => {
             // Successful sign-in, you can access user information here
             const user = result.user;
-            console.log(user);
-            console.log(user.displayName);
-            createAccount(user.uid, user.displayName);
-            setUserID(user.uid);
+            const idToken = await user.getIdToken();
+            const isNewUser = result.additionalUserInfo.isNewUser;
+
+            if (isNewUser || forceRecreate == "true") {
+                createAccount(user.uid, user.displayName, idToken, setJwtToken, rawAccessToken, accessToken, setErrorMessage);
+            } else {
+                signIn(user.uid, idToken, setJwtToken, rawAccessToken, accessToken, setErrorMessage);
+            }
+            //createCookie('userID', user.uid);
             createCookie('userName', user.displayName)
-            createCookie('userID', user.uid);
             createCookie("profileImage", user.photoURL);
             })
             .catch((error) => {
@@ -27,11 +31,14 @@ export function signInWithGoogle(setUserID) {
   };
   
 
-export function createAccount(userID, displayName) {
+export function createAccount(userID, displayName, idToken, setJwtToken, rawAccessToken, accessToken, setErrorMessage) {
     const endpoint = serverURL + 'create-account';
     const data = {
         userID: userID,
-        displayName: displayName
+        displayName: displayName,
+        rawAccessToken: rawAccessToken,
+        accessToken: accessToken,
+        idToken: idToken
     }; // JSON payload
 
     fetch(endpoint, {
@@ -44,16 +51,59 @@ export function createAccount(userID, displayName) {
     })
     .then(response => {
         if (response.ok) {
-        return response.json(); // Parse response JSON if successful
+            return response.json(); // Parse response JSON if successful
         }
         throw new Error('Network response was not ok.');
     })
     .then(data => {
         console.log('Account creation successful:', data);
-        // Handle success response here
+        createCookie("jwtToken", data[0]["token"]);
+        setJwtToken(data[0]["token"])
+        // Set the accessTokens as cookies
+        createCookie("accessToken", accessToken);
+        createCookie("rawAccessToken", rawAccessToken);
     })
     .catch(error => {
         console.error('There was an error creating the account:', error);
+        setErrorMessage("There was an error creating the account");
+        // Handle error here
+    });
+}
+
+export function signIn(userID, idToken, setJwtToken, rawAccessToken, accessToken, setErrorMessage) {
+    const endpoint = serverURL + 'sign-in';
+    const data = {
+        userID: userID,
+        rawAccessToken: rawAccessToken,
+        accessToken: accessToken,
+        idToken: idToken
+    }; // JSON payload
+
+    fetch(endpoint, {
+        method: 'POST',
+        headers: {
+        'Content-Type': 'application/json'
+        // Add any other headers if required
+        },
+        body: JSON.stringify(data) // Convert data to JSON string
+    })
+    .then(response => {
+        if (response.ok) {
+            return response.json(); // Parse response JSON if successful
+        }
+        throw new Error('Network response was not ok.');
+    })
+    .then(data => {
+        console.log('Account creation successful:', data);
+        createCookie("jwtToken", data[0]["token"]);
+        setJwtToken(data[0]["token"])
+        // Set the accessTokens as cookies
+        createCookie("accessToken", accessToken);
+        createCookie("rawAccessToken", rawAccessToken);
+    })
+    .catch(error => {
+        console.error('There was an error creating the account:', error);
+        setErrorMessage("There was an error creating the account");
         // Handle error here
     });
 }