+++ frontend/src/containers/FlashcardOverview/FlashcardOverview.js
@@ -2,8 +2,25 @@ import React from 'react';
 import FlashcardFolder from './FlashcardFolder/FlashcardFolder';
 import FlashcardItem from './FlashcardItem/FlashcardItem';
 import Heading5 from '../../componments/Text/Heading5/Heading5';
+function FlashcardOverview({
+  flashcardData,
+  setMoveFolderDialogueVisible,
+  showDeleteFlashcardConfirmation,
+  setRenameFlashcardSetPopupVisible,
+  setDeleteFolderConfirmationVisible,
+  setRenameFolderPopupVisible,
+  view,
+  selected,
+  setSelected
+}) {
 
-function FlashcardOverview({ flashcardData, setMoveFolderDialogueVisible, view, selected, setSelected }) {
+  function filterFlashcardSet(cardData, title) {
+    /*
+    Remove everything from flashcard data except the flashcard with the title
+    */
+    return cardData[title];
+  }
+
   const renderElement = (element, folderName, path="") => {
     if (element.cards) {
       return (
@@ -11,7 +28,9 @@ function FlashcardOverview({ flashcardData, setMoveFolderDialogueVisible, view, 
           <FlashcardItem
             element={element}
             setMoveFolderDialogueVisible={setMoveFolderDialogueVisible}
-            flashcardData={flashcardData}
+            showDeleteConfirmation={showDeleteFlashcardConfirmation}
+            setRenameFlashcardSetPopupVisible={setRenameFlashcardSetPopupVisible}
+            flashcardData={filterFlashcardSet(flashcardData, element.flashcardName)}
             path={path}
             view={view}
             selected={selected}
@@ -31,16 +50,18 @@ function FlashcardOverview({ flashcardData, setMoveFolderDialogueVisible, view, 
           element={element}
           name={folderName}
           folderKey={folderName + element.flashcardID}
+          showDeleteConfirmation={setDeleteFolderConfirmationVisible}
+          setRenameFolderPopupVisible={setRenameFolderPopupVisible}
+          flashcardData={element}
           child={Object.entries(element).map(([key, value]) => renderElement(value, key, path))}
           view={view}
+          path={path}
         />
       );
     }
   };
   // If flashcardData has no data, return a message to the user
   // If flashcardData has data, it is == ["User has no flashcards"]
-  console.log(flashcardData)
-  console.log(Array.isArray(flashcardData));
   return Array.isArray(flashcardData)
   ? <Heading5 text="You don't have any flashcards yet!"/>
   : <div>{Object.entries(flashcardData).map(([key, value]) => renderElement(value, key))}</div>;